(#131): added ts template unit tests to script, moved type extension to plugin file, minor changes
