--- templates/app-ts/@types/index.d.ts
@@ -1,12 +1,0 @@
-import * as http from 'http'
-import fastify from 'fastify'
-
-declare module 'fastify' {
-  export interface FastifyInstance<
-  HttpServer = http.Server,
-  HttpRequest = http.IncomingMessage,
-  HttpResponse = http.ServerResponse
-> {
-    someSupport(): string;
-  }
-}
