+++ templates/app-ts/services/root.ts
@@ -1,17 +1,7 @@
-import { FastifyInstance, FastifyRequest, FastifyReply } from "fastify"
+import { FastifyInstance } from "fastify"
 
-module.exports = function (fastify: FastifyInstance, opts: any, next: Function) {
-  fastify.get('/', function (request: FastifyRequest, reply: FastifyReply<any>) {
-    reply.send({ root: true })
+module.exports = async function (fastify: FastifyInstance, opts: any) {
+  fastify.get('/', async function (request, reply) {
+    return { root: true }
   })
-
-  next()
 }
-
-// If you prefer async/await, use the following
-//
-// module.exports = async function (fastify, opts) {
-//   fastify.get('/', async function (request, reply) {
-//     return { root: true }
-//   })
-// }
